requirements.txt
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cssselect==1.2.0
cssutils==2.10.2
Flask==3.0.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==5.2.1
MarkupSafe==2.1.5
numpy==2.3.3
pandas==2.3.2
premailer==3.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.31.0
six==1.17.0
tzdata==2025.2
urllib3==2.2.1
Werkzeug==3.0.2
